Skip ChooseAccount Step when configured
This implements reading the IBAN and FireFly III ID from the
config file.
These values will get checked if they exists in the Bank Accounts and in
their Firefly III Account/Instance.
The page will automatically refresh after 2 seconds and the user does
not need to confirm it.
If the check failed an errorpage will be displayed instead.