Enable decoupled TAN submission
This is required for example for TAN mode 923 called pushTAN 2.0 by Sparkasse.
This mode requires a confirmation in the pushTAN app but no TAN to be entered.
The implementation presents the same tan-challenge screen without the text input. Instead of submitting a TAN, the decoupled submission is checked through the fints-library.

Refs bnw/firefly-iii-fints-importer#131