Enable FinTS persistence to stop TAN request for every import
Decoupled TAN modes tend to ask for a TAN in every import run. Using persist on the FinTS instance and loading it from the configuration.json file prevents that.
The string is presented in the done step to be pasted into the configuration file.
If it is present there, it is loaded in the FinTsFactory and used for initialising the FinTS connection.
Refs bnw/firefly-iii-fints-importer#149